% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in license.R
\name{license}
\alias{license_agpl}
\alias{license_apache}
\alias{license_artistic}
\alias{license_cc0}
\alias{license_gpl2}
\alias{license_gpl3}
\alias{license_lgpl2}
\alias{license_lgpl3}
\alias{license_bsd2}
\alias{license_bsd3}
\alias{license_mit}
\title{Add a license to a package}
\usage{
license_agpl(pkgdir = NULL)
license_apache(pkgdir = NULL)
license_artistic(pkgdir = NULL)
license_cc0(pkgdir = NULL)
license_gpl2(pkgdir = NULL)
license_gpl3(pkgdir = NULL)
license_lgpl2(pkgdir = NULL)
license_lgpl3(pkgdir = NULL)
license_bsd2(pkgdir = NULL, year = NULL, copyright_holder = NULL)
license_bsd3(pkgdir = NULL, year = NULL, copyright_holder = NULL, organization = NULL)
license_mit(pkgdir = NULL, year = NULL, copyright_holder = NULL)
}

\arguments{
\item{pkgdir}{The path to the package directory. Defaults to \code{NULL}.}

\item{year}{Year (e.g., 2026). Only used by 'license_bsd2()', 'license_bsd3()' and 'license_mit()'.}

\item{copyright_holder}{Copyright holder (e.g, "package XYZ authors"). Only used by 'license_bsd2()', 'license_bsd3()' and 'license_mit()'.}

\item{organization}{Organization (e.g., "University of Surrey"). Only used by 'license_bsd3()'.}
}

\value{
Invisibly returns 'TRUE' if license was added successfully.
}

\description{
Copy a license file into the package's 'LICENSE.txt', update
 the 'License' field in 'DESCRIPTION', and add 'LICENSE.txt' to
 '.Rbuildignore'. 'license_bsd2()', 'license_bsd3()' and 'license_mit()'
 additionally write the templated 'LICENSE' stub file required by those
 licenses. Available functions: 'license_agpl' (AGPL-3.0), 'license_apache'
 (Apache 2.0), 'license_artistic' (Artistic 2.0), 'license_cc0' (CC0 1.0),
 'license_gpl2' (GPL-2), 'license_gpl3' (GPL-3), 'license_lgpl2' (LGPL-2.1),
 'license_lgpl3' (LGPL-3), 'license_bsd2' (BSD 2-Clause), 'license_bsd3'
 (BSD 3-Clause) and 'license_mit' (MIT).
}

