% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pkg_build}
\alias{pkg_build}
\title{Build R package}
\usage{
pkg_build(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns 'TRUE' if the package was built successfully.
}
\description{
Build an R package into a tar.gz file. The built file will be
created one level up from the package directory.
}
