% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{pkg_clean}
\alias{pkg_clean}
\title{Clean R package binaries}
\usage{
pkg_clean(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns 'TRUE' if binaries were removed successfully.
}
\description{
Remove compiled binaries from the 'src/' directory of an R package.
Only applies to packages with compiled C/C++ code and will remove files with
extensions '.o', '.so', and '.dll'.
}
