% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{pkg_document}
\alias{pkg_document}
\title{Document R package}
\usage{
pkg_document(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns 'TRUE' if the documentation was generated successfully.
}
\description{
Generate documentation for an R package using roxygen2.
}
