% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pkg_load}
\alias{pkg_load}
\title{Load package into the current R session}
\usage{
pkg_load(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns a character vector of loaded R files.
}
\description{
Load all R code and compiled shared libraries from a package directory.
}
