% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in register.R
\name{pkg_register}
\alias{pkg_register}
\title{Register 'C++' Functions}
\usage{
pkg_register(pkgdir = NULL)
}

\arguments{
\item{pkgdir}{The path to the package directory. Defaults to \code{NULL}.}
}

\value{
Invisibly returns 'TRUE' if the documentation was generated successfully.
}

\description{
Register 'C++' functions in an R package that uses 'cpp4r'.
}

