% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydev-package.R
\docType{package}
\name{tinydev-package}
\alias{tinydev}
\alias{tinydev-package}
\title{A Tiny R Package Development Toolkit}
\description{
A minimalistic R package development toolkit that provides
essential functions for creating R packages with ease. 'tinydev' is to
'devtools' what 'tinytest' is to 'testthat', a lightweight alternative that
focuses on simplicity and ease of use.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
