% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tinydev-utils}
\alias{tinydev-utils}
\alias{stop_if_not_package}
\title{Utility functions for tinydev}
\usage{
stop_if_not_package(path)
}
\description{
Internal utility functions for tinydev package.
}
\keyword{internal}
