% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license.R
\name{license_agpl}
\alias{license_agpl}
\alias{license_apache}
\alias{license_artistic}
\alias{license_cc0}
\alias{license_gpl2}
\alias{license_gpl3}
\alias{license_lgpl2}
\alias{license_lgpl3}
\alias{license_bsd2}
\alias{license_bsd3}
\alias{license_mit}
\title{AGPL-3.0 License}
\usage{
license_agpl(pkgdir = ".")

license_apache(pkgdir = ".")

license_artistic(pkgdir = ".")

license_cc0(pkgdir = ".")

license_gpl2(pkgdir = ".")

license_gpl3(pkgdir = ".")

license_lgpl2(pkgdir = ".")

license_lgpl3(pkgdir = ".")

license_bsd2(pkgdir = ".", year = NULL, copyright_holder = NULL)

license_bsd3(
  pkgdir = ".",
  year = NULL,
  copyright_holder = NULL,
  organization = NULL
)

license_mit(pkgdir = ".", year = NULL, copyright_holder = NULL)
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}

\item{year}{Year (e.g., 2026)}

\item{copyright_holder}{Copyright holder (e.g, "package XYZ authors")}

\item{organization}{Organization (e.g., "University of Surrey")}
}
\value{
Invisibly returns 'TRUE' if license was added successfully.
}
\description{
Add AGPL-3.0 license to a package.

Add Apache 2.0 license to a package.

Add Artistic 2.0 license to a package.

Add CC0 1.0 license to a package

Add GPL 2 license to a package

Add GPL 3 license to a package

Add LGPL 2 license to a package

Add LGPL 3 license to a package

Add BSD 2 license to a package plus 'LICENSE' file

Add BSD 3 license to a package plus 'LICENSE' file

Add MIT license to a package plus 'LICENSE' file
}
