% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pkg_build}
\alias{pkg_build}
\title{Build R package}
\usage{
pkg_build(
  pkgdir = ".",
  document = TRUE,
  build_vignettes = TRUE,
  build_manual = TRUE
)
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}

\item{document}{\code{[logical]} Generate documentation before checking.}

\item{build_vignettes}{\code{[logical]} Build vignettes. Defaults to TRUE.}

\item{build_manual}{\code{[logical]} Build manual. Defaults to TRUE.}
}
\value{
Prints the path to the built tar.gz file if successful.
}
\description{
Build an R package into a tar.gz file. The built file will be
created one level up from the package directory.
}
