% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pkg_check}
\alias{pkg_check}
\title{Check R package}
\usage{
pkg_check(pkgdir = ".", cran = TRUE, manual = FALSE, vignettes = FALSE)
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}

\item{cran}{\code{[logical]} Run checks as if on CRAN (passes \code{--as-cran}).}

\item{manual}{\code{[logical]} Build the PDF manual (requires LaTeX).}

\item{vignettes}{\code{[logical]} Build vignettes during check.}
}
\value{
Invisibly returns 'TRUE' if the package was built successfully.
}
\description{
Build and check an R package using R CMD check.
}
