% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in coverage.R
\name{pkg_coverage}
\alias{pkg_coverage}
\title{Check R package coverage}
\usage{
pkg_coverage(pkgdir = NULL, quiet = TRUE, type = "all")
}

\arguments{
\item{pkgdir}{The path to the package directory. Defaults to \code{NULL}.}

\item{type}{The type of coverage to calculate. One of "tests", "examples", "vignettes", or "all". Defaults to "all".}

\item{quiet}{If \code{TRUE}, suppress output from the underlying \code{covr::package_coverage()} function. Defaults to \code{TRUE}.}
}

\value{
Invisibly returns 'TRUE' if the package was built successfully.
}

\description{
Check how much of an R package is tests by unit tests, vignettes and/or examples.
}

