% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{pkg_coverage}
\alias{pkg_coverage}
\title{Check R package coverage}
\usage{
pkg_coverage(pkgdir = ".", quiet = TRUE, type = "all")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}

\item{quiet}{If \code{TRUE}, suppress output from the underlying \code{covr::package_coverage()} function. Defaults to \code{TRUE}.}

\item{type}{The type of coverage to calculate. One of "tests", "examples", "vignettes", or "all". Defaults to "all".}
}
\value{
Invisibly returns 'TRUE' if the package was built successfully.
}
\description{
Check how much of an R package is tests by unit tests, vignettes and/or examples.
}
