% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{pkg_install}
\alias{pkg_install}
\title{Install R package}
\usage{
pkg_install(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns 'TRUE' if the package was installed successfully.
}
\description{
Install an R package from a directory
}
