% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{pkg_template}
\alias{pkg_template}
\title{Start a new project with the cpp4r package template}
\usage{
pkg_template(path = NULL, pkgname = NULL)
}
\arguments{
\item{path}{Path to the new project}

\item{pkgname}{Name of the new package}
}
\value{
Invisibly returns 'TRUE' if the template was copied successfully.
}
\description{
Copy a minimal package template into a new directory.
}
\examples{
# create a new directory
dir <- tempdir()
dir.create(dir)

# copy the package template into the directory
pkg_template(dir, "mynewpkg")
}
