% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{pkg_test}
\alias{pkg_test}
\title{Run tests}
\usage{
pkg_test(pkgdir = ".")
}
\arguments{
\item{pkgdir}{The path to the package directory. Defaults to the current directory.}
}
\value{
Invisibly returns 'TRUE' if tests were run successfully.
}
\description{
Run tests for an R package using tinytest. Tests should be placed
in the 'inst/tinytest' directory of the package. Any R files in that directory
not starting with 'test' (e.g. 'helper.R') are sourced into the global
environment before running the tests, making their definitions available to
all test files.
}
