% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbuildignore.R
\name{rbuildignore_add}
\alias{rbuildignore_add}
\title{Add a pattern to .Rbuildignore}
\usage{
rbuildignore_add(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to add (e.g., "LICENSE.txt" or "^LICENSE\\.txt$").}
}
\value{
Invisibly returns 'TRUE' if the pattern was added successfully or already exists in the file.
}
\description{
Add a pattern to the .Rbuildignore file in the current directory.
If the file does not exist, it will be created. If the pattern already exists in the file, it will not be added again.
}
